+++ lotto_app/app/commands/command_utils.py
@@ -6,6 +6,9 @@ from django.db.models.functions import Cast
 from lotto_app.app.models import Game
 from lotto_app.config import get_from_config
 from lotto_app.constants import COMBINATION_OPTIONS_8_ADD
+from rest_framework.response import Response
+from rest_framework import status
+from lotto_app.app.utils import str_to_list_of_int
 
 
 class Utils8Add():
@@ -133,9 +136,94 @@ class InfoSequence8Add(Utils8Add):
 
 
 class Probabilities8Add(InfoSequence8Add):
+    def _get_probability_objs(self, game_id, steps_back_games, gen_objs):
+        ids = [str(_id) for _id in range(int(game_id), int(game_id) - steps_back_games, -1)]
+        return [obj for obj in gen_objs if obj.game_id in ids]
+
     def get_count_sequences(self, part_consists_of, steps_back_games, limit):
         return {name_seq: sum
                 for name_seq, sum in self.get_all_sequences_in_games(part_consists_of,
                                                                      steps_back_games).items()
                 if sum >= limit
                 }
+
+
+class Probabilities8AddOneNumber():
+
+    def _part_previous(self, ng, previous_id, steps_back_games_previous, gen_probability):
+        return InfoSequence8Add(
+            ng, previous_id, steps_back_games_previous,
+            game_objs=gen_probability._get_probability_objs(previous_id, steps_back_games_previous,
+                                                            gen_probability.game_objs))
+
+    def _part_big(self, ng, big_id, steps_back_games_big, gen_probability):
+        return InfoSequence8Add(
+            ng, big_id, steps_back_games_big,
+            game_objs=gen_probability._get_probability_objs(big_id, steps_back_games_big,
+                                                            gen_probability.game_objs))
+
+    def _list_repeat(self, i_s, steps_back_games):
+        list_repeat = []
+        for str_number, sum in i_s.get_all_sequences_in_games(1, steps_back_games).items():
+            if sum == steps_back_games:
+                list_repeat.extend(str_to_list_of_int(str_number))
+        return list_repeat
+
+    def _get_set_one_numbers_by_big(self, ng, part_big, steps_back_games_small, gen_probability):
+        list_one_numbers = []
+        for obj in part_big.game_objs:
+            i_s = InfoSequence8Add(ng, obj.game_id, steps_back_games_small,
+                                   game_objs=gen_probability._get_probability_objs(obj.game_id,
+                                                                                   steps_back_games_small,
+                                                                                   gen_probability.game_objs))
+            list_one_numbers.extend(self._list_repeat(i_s, steps_back_games_small))
+        return set(list_one_numbers)
+
+    def probability_one_number(self, ng, game_start, how_games,
+                               steps_back_games_previous,
+                               steps_back_games_small,
+                               steps_back_games_big):
+        game_end = game_start - how_games
+
+        probability_one_number = {}
+        gen_probability = Probabilities8Add(
+            ng, game_start,
+            how_games + steps_back_games_previous + steps_back_games_big
+        )
+        gen_obj_end = [_obj for _obj in gen_probability.game_objs][-1]
+        if int(gen_obj_end.game_id) != game_end - steps_back_games_previous - steps_back_games_big:
+            return Response({"error": "there are not enough objects."},
+                            status=status.HTTP_400_BAD_REQUEST)
+
+        set_not_needed_id = set()
+
+        for obj in gen_probability.game_objs:
+            if game_end and int(obj.game_id) > game_end:
+                previous_id = int(obj.game_id)-1
+                part_previous = self._part_previous(ng, previous_id, steps_back_games_previous, gen_probability)
+
+                big_id = previous_id - steps_back_games_previous
+                part_big = self._part_big(ng, big_id, steps_back_games_big, gen_probability)
+
+                set_one_numbers_by_big = self._get_set_one_numbers_by_big(ng,
+                                                                          part_big,
+                                                                          steps_back_games_small,
+                                                                          gen_probability)
+                set_one_numbers_by_previous = set(
+                    self._list_repeat(part_previous,
+                                      len([pp for pp in part_previous.game_objs]))
+                )
+                set_one_numbers = set_one_numbers_by_big & set_one_numbers_by_previous
+                if set_one_numbers_by_big and set_one_numbers_by_previous and set_one_numbers and not (
+                    set_not_needed_id & {game_obj.game_id for game_obj in part_previous.game_objs}
+                ):
+                    probability_one_number[obj.game_id] = {
+                        'obj.numbers': obj.numbers,
+                        'set_one_numbers': set_one_numbers,
+                        'set_one_numbers_by_previous': set_one_numbers_by_previous,
+                        'numbers_have': 1 if set_one_numbers & set(obj.numbers)
+                        else 0
+                    }
+                    if probability_one_number[obj.game_id]['numbers_have']:
+                        set_not_needed_id.add(obj.game_id)
+        return probability_one_number, set_not_needed_id
